Simplify Sequence API, removing a bunch of wrapper methods
This commit drops all the convenience wrappers for __getitem__, __len__,
and so on.  They were simply calling get() and the variable, and then
calling the underlying method.

Let's not waste time adding and maintaining all this.

Also, the API is a bit clearer when you can see that you are actually
calling get(), instead of having it done implicitly.  Especially since we
cannot wrap the most important access method, namely the get().

And for recipe/eval context, it is all wrapped nicely by the EvalWrapper
class.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>